+++ experimental/framework/get-container/src/index.ts
@@ -3,8 +3,8 @@
  * Licensed under the MIT License.
  */
 
-export * from "./routerliciousService";
 export * from "./getSessionStorageContainer";
-export * from "./getTinyliciousContainer";
 export * from "./getContainer";
+export * from "./routerliciousService";
+export * from "./tinyliciousService";
 export { InsecureTinyliciousUrlResolver } from "@fluidframework/tinylicious-driver";
